Add separate entrypoint for testing utils
