+++ src/lib/index.ts
@@ -1,8 +1,4 @@
 export { createSagaMiddleware } from './createSagaMiddleware';
 export { Saga, SagaEnvironment, AnySaga, Task } from './types';
 export { createTypedForEvery, createTypedForLatest } from './createTypedWatchers';
-export { expectSaga } from './testing/expect-saga';
-export { expectBoundFunction } from './testing/expect-bound-function';
-export { call, run, select, spawn } from './testing/mocks';
-export { testSagaWithState, calls, runs, selects, spawns, dispatches } from './stateBasedTestHelper';
 export { default as TimeoutError } from './TimeoutError';
