+++ testing.js
@@ -1,0 +1,1 @@
+export * from './compiled/lib/testing';
